simplify UpstreamFrameSubscriber
